Refactor deploy
